IFile: Use int64_t for file sizes
We can't use an unsigned type since sqlite only offers load/bind
functions for signed 64bits integers
Refs #167

+++ include/medialibrary/IFile.h
@@ -69,7 +69,7 @@ public:
     virtual const std::string& mrl() const = 0;
     virtual Type type() const = 0;
     virtual unsigned int lastModificationDate() const = 0;
-    virtual unsigned int size() const = 0;
+    virtual int64_t size() const = 0;
     virtual bool isRemovable() const = 0;
     ///
     /// \brief isExternal returns true if this stream isn't managed by the medialibrary
