+++ src/filesystem/unix/File.h
@@ -38,11 +38,11 @@ public:
     explicit File( const std::string& filePath, const struct stat& s );
 
     virtual unsigned int lastModificationDate() const override;
-    virtual unsigned int size() const override;
+    virtual int64_t size() const override;
 
 private:
     unsigned int m_lastModificationDate;
-    unsigned int m_size;
+    int64_t m_size;
 };
 
 }
